basic evaluation model is completed with interface

+++ src/main/kotlin/com/finance/stocksimulator/financials/calculations/KeyFiguresCalcs.kt
@@ -70,7 +70,8 @@ class KeyFiguresCalcs() {
        }
 
 //        dividend yield
-        fun dividendYieldPercentage(annualDividendPerShare: Long, stockPrice: Double): Double{
+
+        fun dividendYieldPercentage(annualDividendPerShare: Double, stockPrice: Double): Double{
             return annualDividendPerShare.div(stockPrice).times(100)
         }
